requirements.txt
# Core Data & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Deep Learning
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.0

# Vector Store
faiss-cpu>=1.7.4

# Stock Data
yfinance>=0.2.31

# News Scraping (Yahoo RSS - no API key needed)
feedparser>=6.0.10
beautifulsoup4>=4.12.0
requests>=2.31.0

# Technical Indicators
ta>=0.11.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0

# Testing
pytest>=7.4.0
